requirements.txt
Flask-SQLAlchemy==2.5.1
SQLAlchemy~=1.4.32
PyYAML==6.0
Flask~=2.0.3
pymysql==1.0.2
boto3==1.22.5
sklearn~=0.0
scikit-learn~=1.1.0
pandas~=1.4.2
requests~=2.27.1
bs4~=0.0.1
beautifulsoup4~=4.11.1